Added the `SmartyCachingMiddleware` to the cache application.
The middleware includes an annotation `SmartyCachesTemplate` which takes the name of the template to cache as its argument. If the middleware is installed and `SMARTY_TEMPLATE_CACHING` is enabled the middleware will ask if Smarty it has cached the template and return its content if it does  (thus short-curcuiting the request cycle; the annotated action will not be invoked if Smarty has cached the template).