Added the `ActionFactory` interface, which makes it possible to lazy-load custom actions in the URL configuration. If a request resolves to an `ActionFactory` instance, it calls the factory object's `createAction` method with the request object, the dispatcher, the current resolver, the URL specification and any additional options as arguments.
A simple `ActionFactory` implementation is provided in the `GenericActionFactory` class. It stores the name of an `Action` class when constructed and defers the creation of the action until `createAction` is called (i.e. when a request resolves to the factory instance).

Additionally, the `@Conceal` annotation is added; use it for callable actions which should not be exposed to the dispatch framework.