Added the `redirects` application, which provides a common model for permanent redirects (on a specific `Site` or otherwise).
The `RedirectFallbackMiddleware` can intercept 404s and redirect to a new location if the current request URL is found in the lookup table. If the URL is found and a new URL is specified, the middleware issues a `HTTPResponsePermanentRedirect`. If the new URL is unspecified, the middleware returns a `HTTPResponseGone`. Otherwise, the response object is left untouched.

This application is based on the Django redirects app: http://docs.djangoproject.com/en/dev/ref/contrib/redirects/