Added `SessionNotificationMiddleware`, which adds a simple session messagaging API to the request object.
Call `info`, `error`, `warning` or `success` with a message on the `$request->notifications` object to add notifications for the current request. Retrieve the notifications by iterating over `$request->notifications`. Unless `AUTOCLEAR_SESSION_NOTIFICATIONS` is set to false or the response is a redirect, notifications are automatically deleted after the response is flushed.